dom/l10n/tests/mochitest/document_l10n/test_telemetry.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test DocumentL10n Telemetry</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript">
  "use strict";
  const { Services } = ChromeUtils.import("resource://gre/modules/Services.jsm");
  const { BrowserTestUtils } = ChromeUtils.import(
    "resource://testing-common/BrowserTestUtils.jsm"
  );

  SimpleTest.waitForExplicitFinish();

  function countValues(values) {
    return Object.values(values).reduce((sum, n) => sum + n, 0);
  }

  (async function() {
      let snapshot = Services.telemetry
          .getKeyedHistogramById("L10N_DOCUMENT_INITIAL_TRANSLATION_TIME_US").snapshot();
      // Make sure we collected telemetry from the first window
      is(snapshot.hasOwnProperty("browser_first_window"), true);
      // And we collected only one value
      is(countValues(snapshot.browser_first_window.values), 1);
      // No other window has been recorded yet
      is(snapshot.hasOwnProperty("browser_new_window"), false);
      is(snapshot.hasOwnProperty("about:preferences"), false);

      // Open a new window
      let win = await BrowserTestUtils.openNewBrowserWindow({
        waitForTabURL: "about:blank",
      });

      snapshot = Services.telemetry
          .getKeyedHistogramById("L10N_DOCUMENT_INITIAL_TRANSLATION_TIME_US").snapshot();
      is(countValues(snapshot.browser_first_window.values), 1);
      is(snapshot.hasOwnProperty("browser_new_window"), true);
      is(countValues(snapshot.browser_new_window.values), 1);

      // Open preferences in a new tab
      let tab = BrowserTestUtils.addTab(
          win.gBrowser,
          "about:preferences"
      );
      await BrowserTestUtils.browserLoaded(tab.linkedBrowser);

      snapshot = Services.telemetry
          .getKeyedHistogramById("L10N_DOCUMENT_INITIAL_TRANSLATION_TIME_US").snapshot();

      is(snapshot.hasOwnProperty("about:preferences"), true);
      is(countValues(snapshot["about:preferences"].values), 1);

      await BrowserTestUtils.closeWindow(win);

      SimpleTest.finish();
  })();
  </script>
</head>
<body>
</body>
</html>
